@@ -2,40 +2,69 @@
   "name": "schema-utils",
   "version": "1.0.0",
   "description": "webpack Validation Utils",
-  "main": "src/index.js",
-  "files": [
-    "src"
-  ],
+  "license": "MIT",
+  "repository": "webpack-contrib/schema-utils",
+  "author": "webpack Contrib (https://github.com/webpack-contrib)",
+  "homepage": "https://github.com/webpack-contrib/schema-utils",
+  "bugs": "https://github.com/webpack-contrib/schema-utils/issues",
+  "main": "dist/index.js",
   "engines": {
-    "node": ">= 4"
+    "node": ">= 8.9.0"
   },
   "scripts": {
-    "lint": "eslint --cache src test",
-    "test": "jest --env node --verbose --coverage",
-    "clean": "del-cli coverage",
-    "commits": "commitlint --from $(git rev-list --tags --max-count=1)",
-    "release": "npm run commits && standard-version"
+    "start": "npm run build -- -w",
+    "prebuild": "npm run clean",
+    "build": "cross-env NODE_ENV=production babel src -d dist --ignore \"src/**/*.test.js\" --copy-files",
+    "clean": "del-cli dist",
+    "commitlint": "commitlint --from=master",
+    "lint:prettier": "prettier \"{**/*,*}.{js,json,md,yml,css}\" --list-different",
+    "lint:js": "eslint --cache src test",
+    "lint": "npm-run-all -l -p \"lint:**\"",
+    "prepare": "npm run build",
+    "release": "standard-version",
+    "security": "npm audit",
+    "test:only": "cross-env NODE_ENV=test jest",
+    "test:watch": "cross-env NODE_ENV=test jest --watch",
+    "test:coverage": "cross-env NODE_ENV=test jest --collectCoverageFrom=\"src/**/*.js\" --coverage",
+    "pretest": "npm run lint",
+    "test": "cross-env NODE_ENV=test npm run test:coverage",
+    "defaults": "webpack-defaults"
   },
+  "files": [
+    "dist"
+  ],
   "dependencies": {
     "ajv": "^6.1.0",
     "ajv-errors": "^1.0.0",
     "ajv-keywords": "^3.1.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^7.0.0",
-    "@commitlint/config-conventional": "^7.0.0",
-    "@webpack-contrib/eslint-config-webpack": "^2.0.0",
-    "del-cli": "^1.0.0",
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
+    "@commitlint/cli": "^8.0.0",
+    "@commitlint/config-conventional": "^8.0.0",
+    "@webpack-contrib/defaults": "^5.0.0",
+    "@webpack-contrib/eslint-config-webpack": "^3.0.0",
+    "babel-jest": "^24.8.0",
+    "commitlint-azure-pipelines-cli": "^1.0.2",
+    "cross-env": "^5.2.0",
+    "del": "^4.1.1",
+    "del-cli": "^2.0.0",
     "eslint": "^5.0.0",
+    "eslint-config-prettier": "^4.3.0",
     "eslint-plugin-import": "^2.0.0",
-    "eslint-plugin-prettier": "^2.0.0",
-    "jest": "^21.0.0",
+    "husky": "^2.4.1",
+    "jest": "^24.8.0",
+    "jest-junit": "^6.4.0",
+    "lint-staged": "^8.2.0",
+    "memory-fs": "^0.4.1",
+    "npm-run-all": "^4.1.5",
     "prettier": "^1.0.0",
-    "standard-version": "^4.0.0"
+    "standard-version": "^6.0.1",
+    "webpack": "^4.34.0"
   },
-  "author": "webpack Contrib (https://github.com/webpack-contrib)",
-  "bugs": "https://github.com/webpack-contrib/schema-utils/issues",
-  "homepage": "https://github.com/webpack-contrib/schema-utils",
-  "repository": "https://github.com/webpack-contrib/schema-utils",
-  "license": "MIT"
+  "keywords": [
+    "webpack"
+  ]
 }
