@@ -1,11 +1,4 @@
 module.exports = {
   root: true,
-  plugins: [ 'prettier' ],
-  extends: [ '@webpack-contrib/eslint-config-webpack' ],
-  rules: {
-    'prettier/prettier': [
-      'error',
-      { singleQuote: true, trailingComma: 'es5', arrowParens: 'always' }
-    ]
-  }
+  extends: ['@webpack-contrib/eslint-config-webpack', 'prettier'],
 };
